requirements.txt
pyserial
tqdm
minimalmodbus
